Converting the LispSB to MD-SAL plugin. The phase before writing everything in YANG. TELSDN-125

+++ mappingservice/api/src/main/java/org/opendaylight/lispflowmapping/interfaces/dao/IMappingServiceKey.java
@@ -3,10 +3,13 @@ package org.opendaylight.lispflowmapping.interfaces.dao;
 import org.opendaylight.lispflowmapping.type.lisp.address.LispAddress;
 
 public interface IMappingServiceKey {
-    
+
     LispAddress getEID();
+
     void setEID(LispAddress eID);
-    byte getMask();
+
+    int getMask();
+
     void setMask(byte mask);
 
 }
